+++ projects/rocprofiler-sdk/source/lib/output/output_config.hpp
@@ -68,6 +68,7 @@ struct output_config
     bool                     pftrace_output              = false;
     bool                     otf2_output                 = false;
     bool                     rocpd_output                = false;
+    bool                     rocpd_lz4                   = true;
     bool                     summary_output              = false;
     bool                     kernel_rename               = false;
     bool                     group_by_queue              = false;
@@ -134,6 +135,7 @@ output_config::save(ArchiveT& ar) const
     CFG_SERIALIZE_MEMBER(otf2_output);
     CFG_SERIALIZE_MEMBER(summary_output);
     CFG_SERIALIZE_MEMBER(rocpd_output);
+    CFG_SERIALIZE_MEMBER(rocpd_lz4);
     CFG_SERIALIZE_MEMBER(kernel_rename);
     CFG_SERIALIZE_MEMBER(group_by_queue);
     CFG_SERIALIZE_MEMBER(annotate_args);